+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Overcommit Changelog
 
+## master (unreleased)
+
+* Fix bug where `CHERRY_PICK_HEAD` would be lost when a pre-commit hook failed
+  after attempting to cherry pick a commit with a conflict
+
 ## 0.11.0
 
 * Allow custom arguments to be passed to `ctags` via `IndexTags` post-checkout